feat(dui): propagates changes to all connectors
