+++ ConnectorAutocadCivil/ConnectorAutocadCivil/Entry/SpeckleAutocadCommand.cs
@@ -55,7 +55,7 @@ namespace Speckle.ConnectorAutocadCivil.Entry
     {
       if (MainWindow == null)
       {
-        var viewModel = new MainWindowViewModel(Bindings);
+        var viewModel = new MainViewModel(Bindings);
         MainWindow = new MainWindow
         {
           DataContext = viewModel
