+++ ConnectorCSI/ConnectorCSIShared/cPlugin.cs
@@ -52,7 +52,7 @@ namespace SpeckleConnectorCSI
 
     private static void AppMain(Application app, string[] args)
     {
-      var viewModel = new MainWindowViewModel(Bindings);
+      var viewModel = new MainViewModel(Bindings);
       MainWindow = new MainWindow { DataContext = viewModel };
       MainWindow.Closed += SpeckleWindowClosed;
       MainWindow.Closing += SpeckleWindowClosed;
