+++ ConnectorMicroStation/ConnectorMicroStationOpenShared/Entry/SpeckleMicroStationOpenCommand2.cs
@@ -49,7 +49,7 @@ namespace Speckle.ConnectorMicroStationOpen.Entry
     {
       if (MainWindow == null)
       {
-        var viewModel = new MainWindowViewModel(Bindings);
+        var viewModel = new MainViewModel(Bindings);
         MainWindow = new MainWindow
         {
           DataContext = viewModel
