+++ ConnectorRevit/ConnectorRevit2023/ConnectorRevit2023.csproj
@@ -85,10 +85,10 @@
       <Version>0.10.15</Version>
     </PackageReference>
     <PackageReference Include="Avalonia.Desktop">
-      <Version>0.10.13</Version>
+      <Version>0.10.15</Version>
     </PackageReference>
     <PackageReference Include="Avalonia.ReactiveUI">
-      <Version>0.10.13</Version>
+      <Version>0.10.15</Version>
     </PackageReference>
     <PackageReference Include="MSBuild.AssemblyVersion">
       <Version>1.3.0</Version>
