+++ DesktopUI2/AvaloniaHwndHost/AvaloniaHwndHost.csproj
@@ -14,6 +14,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Avalonia" Version="0.10.0-rc2" />
+    <PackageReference Include="Avalonia" Version="0.10.15" />
   </ItemGroup>
 </Project>
