+++ DesktopUI2/DesktopUI2.WPF/DesktopUI2.WPF.csproj
@@ -8,9 +8,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Avalonia.Direct2D1" Version="0.10.13" />
-    <PackageReference Include="Avalonia.ReactiveUI" Version="0.10.13" />
-    <PackageReference Include="Avalonia.Win32.Interoperability" Version="0.10.13" />
+    <PackageReference Include="Avalonia.Direct2D1" Version="0.10.15" />
+    <PackageReference Include="Avalonia.ReactiveUI" Version="0.10.15" />
   </ItemGroup>
 
   <ItemGroup>
